Engine API V1-V4, hive test compliance (90/129), SNAP sync, ETH/Sepolia
chain support, multi-instance runtime, and all alpha-branch work.

Includes PR #1015 content that was previously only merged to
feature/orbita-architecture.

+++ src/main/scala/com/chipprbots/ethereum/keystore/Wallet.scala
@@ -13,6 +13,6 @@ import com.chipprbots.ethereum.domain.SignedTransactionWithSender
 case class Wallet(address: Address, prvKey: ByteString) {
   lazy val keyPair: AsymmetricCipherKeyPair = keyPairFromPrvKey(prvKey.toArray)
 
-  def signTx(tx: LegacyTransaction, chainId: Option[Byte]): SignedTransactionWithSender =
+  def signTx(tx: LegacyTransaction, chainId: Option[BigInt]): SignedTransactionWithSender =
     SignedTransactionWithSender(SignedTransaction.sign(tx, keyPair, chainId), Address(keyPair))
 }
